# CUDA: Add `fastdiv` to `k_bin_bcast*`, giving 1-3% E2E performance