@JohannesGaessler 2025-09-08
>I was surprised to see we keep ne* in 64 bit precision, but use only the 32 least significant bits in the actual kernel.

The GGML specification for the tensor dimensions is `int64_t`. When work on the CUDA code was first started, `int` was used for everything, largely due to ignorance. You are correct that this could potentially cause issues but unlike e.g. PyTorch ggml is currently not really used by end users directly so the tensor dimensions encountered in practice are very limited. Most kernels are severely I/O bound with negligible register pressure so it would be better to just use 64 bit arguments. For the kernels with non-negligible register pressure more care must be taken but for 32 bit arguments asserts should be added to detect overflows. But as these changes are fairly low-priority no one got around to doing this.